Moves TopLevelNavigationItem into a separate group of related abstractions.
