+++ workspaces/homepage/src/components/navigation/Navigation.js
@@ -3,138 +3,7 @@ import { Link } from 'gatsby'
 import GithubSlugger from 'github-slugger'
 import glamorous from 'glamorous'
 
-const Wrapper = glamorous.div({
-  position: 'relative',
-  display: 'flex',
-  flexFlow: 'row nowrap',
-  alignItems: 'center',
-  backgroundColor: 'black',
-  color: 'white',
-  padding: '8px 13px 9px 22px',
-  ':hover': {
-    cursor: 'pointer'
-  }
-})
-
-const ArrowWrapper = glamorous.div({
-  position: 'absolute',
-  right: '13px'
-})
-
-const ArrowUp = glamorous.div({
-  width: 0,
-  height: 0,
-  borderLeft: '5px solid transparent',
-  borderRight: '5px solid transparent',
-  borderBottom: '8px solid #F486CA'
-})
-
-const ArrowDown = glamorous.div({
-  width: 0,
-  height: 0,
-  borderLeft: '5px solid transparent',
-  borderRight: '5px solid transparent',
-  borderTop: '8px solid #F486CA'
-})
-
-const Text = glamorous.div(({color = 'black'}) => ({
-  fontFamily: 'Roboto Mono, monospace',
-  fontWeight: '300',
-  fontSize: '0.8rem',
-  color: `${color}`
-}))
-
-class TopLevelNavigationItem extends React.Component {
-  state = {
-    folded: true,
-    delta: 0,
-    style: {
-      display: 'flex',
-      flexFlow: 'column nowrap',
-      alignItems: 'flex-start',
-      justifyContent: 'flex-start',
-      maxHeight: 0,
-      overflow: 'hidden',
-      transition: 'max-height 0.2s ease-out'
-    }
-  }
-
-  componentDidUpdate (prevProps) {
-    if (this.props.delta !== prevProps.delta) {
-      console.log('componentDidUpdate', this.props.delta, this.div.scrollHeight)
-      if (!this.state.folded) {
-        this.setState({
-          delta: this.props.delta,
-          style: {
-            ...this.state.style,
-            maxHeight: `${this.div.scrollHeight + this.props.delta}px`
-          }
-        })
-      }
-    }
-  }
-
-  unfold () {
-    if (this.state.folded) {
-      this.setState({
-        folded: false,
-        style: {
-          ...this.state.style,
-          maxHeight: `${this.div.scrollHeight}px`
-        }
-      })
-    } else {
-      this.setState({
-        folded: true,
-        style: {
-          ...this.state.style,
-          maxHeight: 0
-        }
-      })
-    }
-  }
-
-  setDivRef = element => {
-    this.div = element
-  }
-
-  render () {
-    return (
-      <div>
-        <Wrapper onClick={() => this.unfold()}>
-          {this.props.active && <ActiveMarker active={this.props.active} />}
-          <Text color='white'>{this.props.title}</Text>
-          <ArrowWrapper>
-            { this.state.folded ? <ArrowDown /> : <ArrowUp /> }
-          </ArrowWrapper>
-        </Wrapper>
-        <div style={this.state.style} ref={this.setDivRef}>
-          <div style={{backgroundColor: 'white', border: '1px solid #F7F7F7', width: '100%'}}>
-            { this.props.children }
-          </div>
-        </div>
-      </div>
-    )
-  }
-}
-
-const MarkerPositioner = glamorous.div({
-  position: 'absolute',
-  left: '15px'
-})
-
-const Circle = glamorous.div(({radius}) => ({
-  height: radius,
-  width: radius,
-  backgroundColor: '#F486CA',
-  borderRadius: '50%'
-}))
-
-const ActiveMarker = ({active}) => (
-  <MarkerPositioner>
-    <Circle radius='5px' />
-  </MarkerPositioner>
-)
+import { TopLevelNavigationItem } from './top-level-navigation-item/'
 
 class MidLevelNavigationItem extends React.Component {
   state = {
