+++ workspaces/homepage/src/components/navigation/top-level-navigation-item/arrows/ArrowUp.js
@@ -1,0 +1,9 @@
+import glamorous from 'glamorous'
+
+export const ArrowUp = glamorous.div({
+  width: 0,
+  height: 0,
+  borderLeft: '5px solid transparent',
+  borderRight: '5px solid transparent',
+  borderBottom: '8px solid #F486CA'
+})
