+++ workspaces/homepage/src/components/navigation/top-level-navigation-item/arrows/index.js
@@ -1,0 +1,2 @@
+export { ArrowUp } from './ArrowUp'
+export { ArrowDown } from './ArrowDown'
