+++ workspaces/homepage/src/components/navigation/top-level-navigation-item/index.js
@@ -1,0 +1,1 @@
+export { TopLevelNavigationItem } from './TopLevelNavigationItem'
